% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{exampleData}
\alias{exampleData}
\title{Example response time data set}
\format{A data frame with 20518 rows and 4 variables:
\describe{
    \item{participant}{participant identification number}
    \item{condition}{the experimental condition (2 in this example)}
    \item{rt}{response time, coded in milliseconds}
    \item{accuracy}{accuracy of the response; 1 = correct, 0 = error}

    }}
\usage{
exampleData
}
\description{
An example data set containing multiple participants' data for a response
time study involving two experimental conditions. The data set also includes
This is a synthetic data set and has no theoretical basis.
}
\keyword{datasets}

