% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{linearInterpolation}
\alias{linearInterpolation}
\title{SDs used for the recursive / moving criterion trimming methods}
\format{A data frame with 97 rows and 3 columns:
\describe{
    \item{sampleSize}{Sample size of the data set being passed}
    \item{nonRecursive}{The standard deviation to use as the criterion for
    the nonRecursive function}
    \item{modifiedRecursive}{The standard deviation to use as the criterion
    for the modifiedRecursive function}

    }}
\usage{
linearInterpolation
}
\description{
A data frame containing the SDs used for each sample size as trimming
criterion for the nonRecursive function and the modifiedRecursive function
}
\keyword{datasets}

