% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/perCell.R
\name{perCell}
\alias{perCell}
\title{Per Cell SD RT trimming}
\usage{
perCell(data, minRT, sd, omitErrors = TRUE, returnType = "raw",
  seconds = FALSE)
}
\arguments{
\item{data}{A data frame. It must contain columns named "participant",
"condition", "rt", and "accuracy". The RT can be in seconds
(e.g., 0.654) or milliseconds (e.g., 654). Typically, "condition" will
consist of strings. "accuracy" must be 1 for correct and 0 for error
responses.}

\item{minRT}{The lower criteria for acceptable response time. Must be in
the same form as rt column in data frame (e.g., in seconds OR milliseconds).
All RTs below this value are removed before proceeding with SD trimming.}

\item{sd}{The upper criteria for standard deviation cut-off.}

\item{omitErrors}{If set to TRUE, error trials will be removed before
conducting trimming procedure. Final data returned will not be influenced
by errors in this case.}

\item{returnType}{Request nature of returned data. "raw" returns trial-
level data excluding trimmed data; "mean" returns mean response times per
participant for each experimental condition identified; "median" returns
median response times per participant for each experimental condition
identified.}

\item{seconds}{If set to TRUE, the response time in the data frame is
in seconds; retain as FALSE if your data is in millisecond form.}
}
\description{
\code{perCell} takes a data frame of RT data and returns trimmed rt data
that fall below a set standard deviation above the each condition's mean
(e.g., trimming all RTs slower than 2SDs of each condition's mean). This
function uses the group mean to calculate the cutoff criteria.
}
\details{
By passing a data frame containing raw response time data, together with
trimming criteria, the function will return trimmed data, either in the form
of trial-level data or in the form of means/medians.
}
\examples{
To do
}

